requirements.txt
ttkbootstrap
opencv-python
pyzbar
Pillow